refactor: Isolate Google Contacts logic into separate files
This commit further modularizes the codebase by separating
Google Contacts integration logic from the more general autoreply
handlers and types.

Key changes:

- **`google_contacts_types.go`:**
  - New file created.
  - Contains struct definitions specifically for Google Contacts API
    interaction and the `/autoreply/contactgroup...` endpoints.
    These include `AuthTokenRequest`, `ContactGroupRequest`,
    `ContactGroupDeleteRequest`, and all `Google...` API response
    parsing structs (e.g., `GooglePerson`, `GoogleApiError`).
    These were moved from `autoreply_types.go`.

- **`google_contacts_integration.go`:**
  - New file created.
  - Contains handler functions for Google Contacts:
    `SetGoogleContactsAuthToken`, `AddContactGroupToMode`,
    `DeleteContactGroupFromMode`.
  - Helper functions specific to this integration:
    `fetchContactsFromGoogleGroupFunc` (and its implementation) and
    `normalizePhoneNumber` also moved here.
    These were moved from `autoreply_handlers.go`.

- **`autoreply_types.go`:**
  - Cleaned up by removing the Google Contacts specific structs. Now
    primarily contains types for mode management and simple autoreply.

- **`autoreply_handlers.go`:**
  - Cleaned up by removing the Google Contacts specific handlers and
    helpers. Now contains handlers for mode management and simple
    autoreply, and the `isValidModeName` helper.

- **`autoreply_routes.go`:**
  - No changes were needed in this file, as the method signatures
    on the `*server` type remain consistent, and the router correctly
    locates them regardless of the source file within `package main`.

All new and modified files maintain `package main` and have updated
imports as necessary. This refactoring enhances modularity and aims
to reduce potential merge conflicts.
