+++ routes.go
@@ -101,23 +101,8 @@ func (s *server) routes() {
 	s.router.Handle("/chat/send/poll", c.Then(s.SendPoll())).Methods("POST")
 	s.router.Handle("/chat/send/edit", c.Then(s.SendEditMessage())).Methods("POST")
 
-	s.router.Handle("/chat/autoreply", c.Then(s.AddAutoReply())).Methods("POST")
-	s.router.Handle("/chat/autoreply", c.Then(s.DeleteAutoReply())).Methods("DELETE")
-	s.router.Handle("/chat/autoreply", c.Then(s.GetAutoReplies())).Methods("GET")
-
-	// Autoreply Mode Routes (formerly /mode/)
-	s.router.Handle("/autoreply/mode", c.Then(s.AddModeAutoreply())).Methods("POST")
-	s.router.Handle("/autoreply/mode", c.Then(s.DeleteModeAutoreply())).Methods("DELETE")
-	s.router.Handle("/autoreply/mode", c.Then(s.GetModeAutoreplies())).Methods("GET")
-	s.router.Handle("/autoreply/enablemode", c.Then(s.EnableMode())).Methods("POST")
-	s.router.Handle("/autoreply/disablemode", c.Then(s.DisableMode())).Methods("POST")
-	s.router.Handle("/autoreply/currentmode", c.Then(s.GetCurrentMode())).Methods("GET")
-	s.router.Handle("/autoreply/clearmode", c.Then(s.ClearModes())).Methods("POST")
-
-	// Autoreply Contact Group Routes
-	s.router.Handle("/autoreply/contactgroupauth", c.Then(s.SetGoogleContactsAuthToken())).Methods("POST")
-	s.router.Handle("/autoreply/contactgroup", c.Then(s.AddContactGroupToMode())).Methods("POST")
-	s.router.Handle("/autoreply/contactgroup", c.Then(s.DeleteContactGroupFromMode())).Methods("DELETE")
+	// Register all autoreply related routes
+	registerAutoreplyRoutes(s, s.router, c)
 
 	s.router.Handle("/user/presence", c.Then(s.SendPresence())).Methods("POST")
 	s.router.Handle("/user/info", c.Then(s.GetUser())).Methods("POST")
